Déplacer le code de l'outil CLI vers du code serveur pour générer la déclaration GeoMCE
